worker/manager simplification
- split up pause/unpause channel from worker commands
- add modules for manager/worker functions
- add new ManagerCommand to handle scheduling new tasks for workers
