+++ crates/spyglass/src/task.rs
@@ -1,39 +1,41 @@
 use notify::event::ModifyKind;
 use notify::{EventKind, RecursiveMode, Watcher};
 use tokio::sync::{broadcast, mpsc};
-use url::Url;
 
-use crate::pipeline::PipelineCommand;
-use entities::models::{crawl_queue, indexed_document};
-use entities::sea_orm::prelude::*;
-use entities::sea_orm::{ColumnTrait, EntityTrait, QueryFilter, Set};
-use shared::config::{Config, LensConfig};
+use shared::config::Config;
 
 use crate::crawler::Crawler;
-use crate::search::{
-    lens::{load_lenses, read_lenses},
-    Searcher,
-};
+use crate::search::lens::{load_lenses, read_lenses};
 use crate::state::AppState;
 
-#[derive(Debug, Clone)]
-pub enum TaskType {
-    Initialize,
-    Crawl { id: i64 },
-    Tag,
-}
+mod manager;
+mod worker;
 
 #[derive(Debug, Clone)]
 pub struct CrawlTask {
     pub id: i64,
 }
 
-/// Used to control the scheduler
+/// Tell the manager to schedule some tasks
+pub enum ManagerCommand {
+    Collect,
+    CheckForJobs,
+}
+
+/// Send tasks to the worker
 #[derive(Clone, Debug)]
-pub enum Command {
-    Execute(TaskType),
-    PauseCrawler,
-    RunCrawler,
+pub enum WorkerCommand {
+    // Fetch, parses, & indexes a URI
+    // TODO: Split this up so that this work can be spread out.
+    Crawl { id: i64 },
+    // Applies tag information to an URI
+    Tag,
+}
+
+#[derive(Clone, Debug)]
+pub enum AppPause {
+    Pause,
+    Run,
 }
 
 #[derive(Clone, Debug)]
@@ -41,215 +43,50 @@ pub enum AppShutdown {
     Now,
 }
 
-/// Manages the crawl queue
+/// Manages the worker pool, scheduling tasks based on type/priority/etc.
 #[tracing::instrument(skip_all)]
 pub async fn manager_task(
     state: AppState,
-    queue: mpsc::Sender<Command>,
+    queue: mpsc::Sender<WorkerCommand>,
+    manager_cmd_tx: mpsc::UnboundedSender<ManagerCommand>,
+    mut manager_cmd_rx: mpsc::UnboundedReceiver<ManagerCommand>,
     mut shutdown_rx: broadcast::Receiver<AppShutdown>,
 ) {
     log::info!("manager started");
-
+    let mut queue_check_interval = tokio::time::interval(std::time::Duration::from_millis(100));
     loop {
-        let mut prioritized_domains: Vec<String> = Vec::new();
-        let mut prioritized_prefixes: Vec<String> = Vec::new();
-
-        for entry in state.lenses.iter() {
-            let value = entry.value();
-            if value.pipeline.is_none() {
-                prioritized_domains.extend(value.domains.clone());
-                prioritized_prefixes.extend(value.urls.clone());
+        tokio::select! {
+            cmd = manager_cmd_rx.recv() => {
+                if let Some(cmd) = cmd {
+                    match cmd {
+                        ManagerCommand::Collect => {
+                            log::debug!("collecting uris for crawl queue");
+                        }
+                        ManagerCommand::CheckForJobs => {
+                            log::debug!("checking for new jobs");
+                            manager::check_for_jobs(&state, &queue).await
+                        }
+                    }
+                }
             }
-        }
-
-        let next_url = tokio::select! {
-            res = crawl_queue::dequeue(
-                &state.db,
-                state.user_settings.clone(),
-                &prioritized_domains,
-                &prioritized_prefixes,
-            ) => res,
+            _ = queue_check_interval.tick() => {
+                if let Err(err) = manager_cmd_tx.send(ManagerCommand::CheckForJobs) {
+                    log::error!("Unable to send manager command: {}", err.to_string());
+                }
+            }
             _ = shutdown_rx.recv() => {
                 log::info!("🛑 Shutting down manager");
                 return;
             }
         };
-
-        match next_url {
-            Err(err) => log::error!("Unable to dequeue: {}", err),
-            Ok(Some(task)) => {
-                match &task.pipeline {
-                    Some(pipeline) => {
-                        let mut pipeline_tx = state.pipeline_cmd_tx.lock().await;
-                        match &mut *pipeline_tx {
-                            Some(pipeline_tx) => {
-                                log::debug!("Sending crawl task to pipeline");
-                                let cmd = PipelineCommand::ProcessUrl(
-                                    pipeline.clone(),
-                                    CrawlTask { id: task.id },
-                                );
-                                if let Err(err) = pipeline_tx.send(cmd).await {
-                                    log::error!("Unable to send crawl task to pipeline {:?}", err);
-                                }
-                            }
-                            None => {
-                                log::error!(
-                                    "Unable to send crawl task to pipeline, no queue found"
-                                );
-                            }
-                        }
-                    }
-                    None => {
-                        // Send to worker
-                        let cmd = Command::Execute(TaskType::Crawl { id: task.id });
-                        if queue.send(cmd).await.is_err() {
-                            log::error!("unable to send command to worker");
-                        }
-                    }
-                }
-            }
-            // Nothing to do... sleepy time!
-            Ok(None) => tokio::time::sleep(tokio::time::Duration::from_secs(5)).await,
-        }
-
-        // Wait a little before we dequeue another URL
-        tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-    }
-}
-
-#[tracing::instrument(skip(state, crawler))]
-async fn _handle_fetch(state: AppState, crawler: Crawler, task: CrawlTask) {
-    let result = crawler.fetch_by_job(&state.db, task.id, true).await;
-
-    match result {
-        Ok(Some(crawl_result)) => {
-            // Update job status
-            // We consider 400s complete in this case since we manage to hit the server
-            // successfully but nothing useful was returned.
-            let cq_status = if crawl_result.is_success() || crawl_result.is_bad_request() {
-                crawl_queue::CrawlStatus::Completed
-            } else {
-                crawl_queue::CrawlStatus::Failed
-            };
-
-            let _ = crawl_queue::mark_done(&state.db, task.id, cq_status).await;
-
-            // Add all valid, non-duplicate, non-indexed links found to crawl queue
-            let to_enqueue: Vec<String> = crawl_result.links.into_iter().collect();
-
-            // Collect all lenses that do not
-            let lenses: Vec<LensConfig> = state
-                .lenses
-                .iter()
-                .filter(|entry| entry.value().pipeline.is_none())
-                .map(|entry| entry.value().clone())
-                .collect();
-
-            if let Err(err) = crawl_queue::enqueue_all(
-                &state.db,
-                &to_enqueue,
-                &lenses,
-                &state.user_settings,
-                &Default::default(),
-                Option::None,
-            )
-            .await
-            {
-                log::error!("error enqueuing all: {}", err);
-            }
-
-            // Only add valid urls
-            // if added.is_none() || added.unwrap() == crawl_queue::SkipReason::Duplicate {
-            //     link::save_link(&state.db, &crawl_result.url, link)
-            //         .await
-            //         .unwrap();
-            // }
-
-            // Add / update search index w/ crawl result.
-            if let Some(content) = crawl_result.content {
-                let url = Url::parse(&crawl_result.url).expect("Invalid crawl URL");
-                let url_host = url.host_str().expect("Invalid URL host");
-
-                let existing = indexed_document::Entity::find()
-                    .filter(indexed_document::Column::Url.eq(url.as_str()))
-                    .one(&state.db)
-                    .await
-                    .unwrap_or_default();
-
-                // Delete old document, if any.
-                if let Some(doc) = &existing {
-                    if let Ok(mut index_writer) = state.index.writer.lock() {
-                        let _ = Searcher::remove_from_index(&mut index_writer, &doc.doc_id);
-                    }
-                }
-
-                // Add document to index
-                let doc_id: Option<String> = {
-                    if let Ok(mut index_writer) = state.index.writer.lock() {
-                        match Searcher::add_document(
-                            &mut index_writer,
-                            &crawl_result.title.unwrap_or_default(),
-                            &crawl_result.description.unwrap_or_default(),
-                            url_host,
-                            url.as_str(),
-                            &content,
-                            &crawl_result.raw.unwrap_or_default(),
-                        ) {
-                            Ok(new_doc_id) => Some(new_doc_id),
-                            _ => None,
-                        }
-                    } else {
-                        None
-                    }
-                };
-
-                if let Some(doc_id) = doc_id {
-                    // Update/create index reference in our database
-                    let indexed = if let Some(doc) = existing {
-                        let mut update: indexed_document::ActiveModel = doc.into();
-                        update.doc_id = Set(doc_id);
-                        update
-                    } else {
-                        indexed_document::ActiveModel {
-                            domain: Set(url_host.to_string()),
-                            url: Set(url.as_str().to_string()),
-                            doc_id: Set(doc_id),
-                            ..Default::default()
-                        }
-                    };
-
-                    if let Err(e) = indexed.save(&state.db).await {
-                        log::error!("Unable to save document: {}", e);
-                    }
-                }
-            }
-        }
-        Ok(None) => {
-            // Failed to grab robots.txt or crawling is not allowed
-            if let Err(e) =
-                crawl_queue::mark_done(&state.db, task.id, crawl_queue::CrawlStatus::Completed)
-                    .await
-            {
-                log::error!("Unable to mark task as finished: {}", e);
-            }
-        }
-        Err(err) => {
-            log::error!("Unable to crawl id: {} - {:?}", task.id, err);
-            // mark crawl as failed
-            if let Err(e) =
-                crawl_queue::mark_done(&state.db, task.id, crawl_queue::CrawlStatus::Failed).await
-            {
-                log::error!("Unable to mark task as failed: {}", e);
-            }
-        }
     }
 }
 
 /// Grabs a task
 pub async fn worker_task(
     state: AppState,
-    mut queue: mpsc::Receiver<Command>,
-    mut crawler_cmd: broadcast::Receiver<Command>,
+    mut queue: mpsc::Receiver<WorkerCommand>,
+    mut pause_rx: broadcast::Receiver<AppPause>,
     mut shutdown_rx: broadcast::Receiver<AppShutdown>,
 ) {
     log::info!("worker started");
@@ -261,8 +98,8 @@ pub async fn worker_task(
         // infinite loop
         if is_paused {
             tokio::select! {
-                res = crawler_cmd.recv() => {
-                    if let Ok(Command::RunCrawler) = res {
+                res = pause_rx.recv() => {
+                    if let Ok(AppPause::Run) = res {
                         is_paused = false;
                     }
                 },
@@ -278,9 +115,16 @@ pub async fn worker_task(
 
         let next_cmd = tokio::select! {
             res = queue.recv() => res,
-            res = crawler_cmd.recv() => res.ok(),
+            res = pause_rx.recv() => {
+                if let Ok(AppPause::Pause) = res {
+                    is_paused = true;
+                }
+
+                None
+            },
             _ = shutdown_rx.recv() => {
                 log::info!("🛑 Shutting down worker");
+                queue.close();
                 return;
             }
         };
@@ -288,19 +132,14 @@ pub async fn worker_task(
         if let Some(cmd) = next_cmd {
             log::debug!("handling command: {:?}", cmd);
             match cmd {
-                Command::PauseCrawler => is_paused = true,
-                Command::RunCrawler => is_paused = false,
-                Command::Execute(task) => match task {
-                    TaskType::Initialize => {}
-                    TaskType::Crawl { id } => {
-                        tokio::spawn(_handle_fetch(
-                            state.clone(),
-                            crawler.clone(),
-                            CrawlTask { id },
-                        ));
-                    }
-                    TaskType::Tag => {}
-                },
+                WorkerCommand::Crawl { id } => {
+                    tokio::spawn(worker::handle_fetch(
+                        state.clone(),
+                        crawler.clone(),
+                        CrawlTask { id },
+                    ));
+                }
+                WorkerCommand::Tag => {}
             }
         }
     }
@@ -310,7 +149,7 @@ pub async fn worker_task(
 pub async fn lens_watcher(
     state: AppState,
     config: Config,
-    mut crawler_cmd: broadcast::Receiver<Command>,
+    mut pause_rx: broadcast::Receiver<AppPause>,
     mut shutdown_rx: broadcast::Receiver<AppShutdown>,
 ) {
     log::info!("👀 lens watcher started");
@@ -332,17 +171,17 @@ pub async fn lens_watcher(
     load_lenses(state.clone()).await;
 
     loop {
+        // Run w/ a select on the shutdown signal otherwise we're stuck in an
+        // infinite loop
         if is_paused {
             tokio::select! {
-                res = crawler_cmd.recv() => {
-                    match res {
-                        Ok(Command::PauseCrawler) => is_paused = true,
-                        Ok(Command::RunCrawler) => is_paused = false,
-                        _ => {}
+                res = pause_rx.recv() => {
+                    if let Ok(AppPause::Run) = res {
+                        is_paused = false;
                     }
                 },
                 _ = shutdown_rx.recv() => {
-                    log::info!("🛑 Shutting down worker");
+                    log::info!("🛑 Shutting down lens watcher");
                     return;
                 }
             };
@@ -353,6 +192,13 @@ pub async fn lens_watcher(
 
         let event = tokio::select! {
             res = rx.recv() => res,
+            res = pause_rx.recv() => {
+                if let Ok(AppPause::Pause) = res {
+                    is_paused = true;
+                }
+
+                None
+            },
             _ = shutdown_rx.recv() => {
                 log::info!("🛑 Shutting down lens watcher");
                 return;
